fix (CheckmateDetector.java) corrected error for int to byte lossy conversion while assiging 6 or -6 or 1 or 0 to byte varibale

+++ TMChessServer/src/main/java/com/thinking/machines/chess/server/logic/CheckmateDetector.java
@@ -164,7 +164,7 @@ public static boolean detectCheckmate(byte [][]board,byte color)
 {
 KingCastling kingCastling=new KingCastling();
 kingCastling.checkCastling=false;
-byte kingPiece=(color>0)?6:-6;
+byte kingPiece=(byte)((color>0)?6:-6);
 int kingRowIndex=0;
 int kingColumnIndex=0;
 for(int e=0;e<8;e++)
